gdm was updated to new paradigm

+++ components/gdm_file/gdm_file.cpp
@@ -12,7 +12,8 @@ void delete_key_word(std::string & input){
     }
 }
 
-GdmFile::GdmFile(){
+GdmFile::GdmFile(std::ostream & log_stream):log_stream(log_stream){
+    log_stream<<"Opening main project file..."<<std::endl;
     file.open("project.gdm");
     if(not file.is_open()){
         std::cerr<<"There is no main project file."<<std::endl;
@@ -22,7 +23,7 @@ GdmFile::GdmFile(){
     path = std::filesystem::absolute(std::filesystem::path("project.gdm"));
 }
 
-GdmFile::GdmFile(std::string in_path){
+GdmFile::GdmFile(std::string in_path, std::ostream & log_stream):log_stream(log_stream){
     std::string curr_path = in_path + "component.gdm";
     file.open(curr_path);
     if(not file.is_open()){
@@ -46,7 +47,9 @@ bool GdmFile::garbage_string_skip(){
             file.seekp(position);
             break;
         }
-        if (safety_counter>100) break;
+        if (position==-1) {
+            return false;
+        }
     }
     return true;
 }
@@ -83,8 +86,22 @@ bool GdmFile::garbage_string_skip(){
                         // isforce);
 // }
 
-std::string GdmFile::take_link(std::string & in){
-
+/**
+ * Возвращает ДА, если ссылка с полным путем
+ * Возвращает НЕТ, если ссылка короткая
+*/
+bool GdmFile::take_link(std::string & in, std::string & link){
+    link = in.substr(0, in.find_first_of(" "));
+    in.erase(0, in.find_first_of(" "));
+    in.erase(0, in.find_first_not_of(" "));
+    if (link.find("//")==0){ // случай, когда указан полный путь
+        link.erase(0,2); 
+        return true;
+    }
+    else {
+        if (link.find_first_of("/")==0) link.erase(0,1);
+        return false;
+    }
 }
 
 void GdmFile::take_branch_and_commit(std::string & in, std::string & branch, std::string & commit){
@@ -101,7 +118,7 @@ void GdmFile::take_branch_and_commit(std::string & in, std::string & branch, std
         in.erase(0,branch.size()+commit.size()+1); // стерли из считанной строки ветку и комит
         return;
     }
-    if(slash_pos==space_pos==std::string::npos){    // Если после ветки совсем ничего нет
+    if((slash_pos==space_pos)&&(space_pos==std::string::npos)){    // Если после ветки совсем ничего нет
         branch = in;
         commit = "HEAD";
         in.erase(0, in.size());
@@ -125,24 +142,39 @@ void GdmFile::take_branch_and_commit(std::string & in, std::string & branch, std
     }
 }
 
+bool GdmFile::check_force(std::string & in){
+    if (path!=std::filesystem::absolute(std::filesystem::path("project.gdm"))){  // force может быть только в корне
+        return false;
+    }
+    if (in.size()<5) return false; 
+    in.erase(0, in.find_first_not_of(" "));
+    std::transform(in.begin(), in.end(), in.begin(), toupper);
+    if (in.find("FORCE")==0){
+        return true;
+    } else return false;
+}
+
 Component GdmFile::get_comp(){
     if (file.eof()){
         return Component();
     }
-    garbage_string_skip(); // пропускаем ненужные строки
+    if (!garbage_string_skip()) return Component(); // пропускаем ненужные строки
     std::string read_line;
     getline(file, read_line);
-        // std::cout<<"read string: "<<read_line<<std::endl;
-    // delete_first_spaces(read_line); //вероятно, эту строку можно удалять
+        log_stream<<"________________"<<std::endl;
+        log_stream<<"read string: "<<read_line<<std::endl;
+        log_stream<<"was parsing as: "<<std::endl;
+        
     delete_key_word(read_line);
-        // std::cout<<"updt string: "<<read_line<<std::endl;
-    std::string root_folder = "NONE";
-    std::string root_rms = "NONE";
-    std::string link = take_link(read_line);
+    std::string link;
+    bool is_full_path =  take_link(read_line, link);
     std::string branch;
     std::string commit;
     take_branch_and_commit(read_line, branch, commit);
-    
+    bool is_force = check_force(read_line);
+    Component temp (_root_rms, _root_folder, link, path, is_full_path, branch, commit, is_force);
+    temp.print_for_log(log_stream);
+    return temp;
 }
 GdmFile::~GdmFile(){
